Group drop behaviours into a single composite behaviour
